Test 'nullable list'
