+++ generator/test/generator_test_io/nullable_list.dart
@@ -1,0 +1,11 @@
+import 'package:realm_common/realm_common.dart';
+
+//part 'nullable_list.g.dart';
+
+@RealmModel()
+class _Bad {
+  @PrimaryKey()
+  late int id;
+
+  List<int>? wrong;
+}
