+++ generator/test/generator_test_io/nullable_list.log
@@ -1,0 +1,11 @@
+Realm collections cannot be nullable
+
+in: asset:pkg/test/generator_test_io/nullable_list.dart:10:3
+    ╷
+5   │ @RealmModel()
+6   │ class _Bad {
+    │       ━━━━ in realm model for 'Bad'
+... │
+10  │   List<int>? wrong;
+    │   ^^^^^^^^^^ is nullable
+    ╵
